+++ test/Feature/CleanRegex/PatternBuilder/builder/PatternBuilderTest.php
@@ -18,7 +18,7 @@ class PatternBuilderTest extends TestCase
         $pattern = $pattern->delimited();
 
         // then
-        $this->assertSame('#You/her, (are|is) real\?\ \(or\ are\ you\ not\ real\?\) (you|her)#', $pattern);
+        $this->assertSame('#You/her, (are|is) real\?\ \(or\ are\ you\ not\ real\?\) (you|her)#uXSD', $pattern);
     }
 
     /**
@@ -35,7 +35,7 @@ class PatternBuilderTest extends TestCase
         $pattern = $pattern->delimited();
 
         // then
-        $this->assertSame('#You/her, (are|is) real\?\ \(or\ are\ you\ not\ real\?\) (you|her)#', $pattern);
+        $this->assertSame('#You/her, (are|is) real\?\ \(or\ are\ you\ not\ real\?\) (you|her)#uXSD', $pattern);
     }
 
     /**
@@ -52,6 +52,21 @@ class PatternBuilderTest extends TestCase
         $pattern = $pattern->delimited();
 
         // then
-        $this->assertSame('#You/her, (are|is) real\?\ \(or\ are\ you\ not\ real\?\) (you|her)#', $pattern);
+        $this->assertSame('#You/her, (are|is) real\?\ \(or\ are\ you\ not\ real\?\) (you|her)#uXSD', $pattern);
+    }
+
+    /**
+     * @test
+     */
+    public function shouldBuild_template_bind_DefaultFlags()
+    {
+        // given
+        $pattern = PatternBuilder::builder()->template('%You/her, \s (her)%')->bind([]);
+
+        // when
+        $pattern = $pattern->delimited();
+
+        // then
+        $this->assertSame('#%You/her, \s (her)%#uXSD', $pattern);
     }
 }